+++ zdt.zig
@@ -14,8 +14,8 @@ pub const ZdtError = @import("./lib/errors.zig").ZdtError;
 
 // make sure 'internal' tests are also executed:
 const calendar = @import("./lib/calendar.zig");
-const psx = @import("./lib/posixtz.zig");
-const string = @import("./lib/string.zig");
+const posix = @import("./lib/posixtz.zig");
+const str = @import("./lib/string.zig");
 const tzif = @import("./lib/tzif.zig");
 
 test {
@@ -24,7 +24,7 @@ test {
     _ = Timezone;
     _ = Duration;
     _ = calendar;
-    _ = psx;
-    _ = string;
+    _ = posix;
+    _ = str;
     _ = tzif;
 }